Refactored html and main to add inheritance and ease maintenance
