+++ templates/about.html
@@ -1,0 +1,8 @@
+{% extends "layout.html" %}
+{% block body %}
+<div class="container">
+<title>About Us </title>
+
+<p>Project created by Ben Schmoker <a href='http://github.com/bschmoker'>Follow him on Github</a> </p>
+</div>
+{% endblock %}
