+++ templates/breach.html
@@ -1,0 +1,76 @@
+{% extends "layout.html" %}
+{% block body %}
+
+<div class="container">
+
+{% if breach_id %}
+    <title>Success </title>
+    <h1>Top men are on the case.
+    </h1>
+
+    <h2>
+        Refer to
+        <a href = {{url_for('breach_results', breach_id=breach_id ,_external = True)}} > this link</a>
+        to track the status of your submission
+    </h2>
+
+{% else %}
+
+    <title>Breach Submission </title>
+    <form method="POST" action="/breach/new" >
+        <p>
+            First, what is your name? <input type="text" name="submitter" required> </input>
+        </p>
+
+        <p>
+            How would you describe what happened? </br> <textarea maxlength = 100 rows="7" cols = "30" name="description" required> </textarea>
+        </p>
+
+        <p>
+            Who did it happen to?<input type="text" name="organization"required>
+        </p>
+
+        <p>
+    <!--# TODO handle if blank, default empty string or None?  -->
+        What did they take (if anything) </br> <input type="text" name="asset" > </input>
+        </p>
+
+        <p>
+            What kind of damage did it cause?
+             <select name="damage">
+                <option value="Brand or Image Degradation">It damaged their image </option>
+                <option value="Loss of Competitive Advantage">They lost business to rivals </option>
+                <option value="Data Breach or Compromise" selected> Their data was stolen</option>
+                <option value="Regulatory, Compliance or Legal Impact">They might get sued</option>
+                <option value="Financial Loss"> They lost money</option>
+                <option value="Degradation of Service"> Their servers went down</option>
+             </select>
+        </p>
+
+        <p> When did it happen? 
+            <input type="date" name="timestamp" required> </input>
+        </p>
+
+        <p>
+            How confident is your assertion?
+             <select name="confidence">
+                <option value=High>Very Confident<option>
+                <option selected value=Medium>Sorta Confident</option>
+                <option value=Low>Not Confident</option>
+                <option value=Unknown>Just a rumor</option>
+             </select>
+        </p>
+
+        <p>
+            Is this information sensitive?
+            <input type="hidden" value="False" name="sensitive" />
+            <input type="checkbox" name="sensitive" value="True" checked > (select if true)</input>
+        </p>
+
+        <input type="submit" value="Submit">
+    </form>
+
+{% endif %}
+
+</div><!-- /.container -->
+{% endblock %}
