+++ templates/breachList.html
@@ -1,0 +1,42 @@
+{% extends "layout.html" %}
+{% block body %}
+<div class="container">
+
+{% if show_all %}
+<title>Breach Details </title>
+<table width = "500px">
+<tr>
+    <th> ID </th>
+    <th> Timestamp </th>
+    <th> Type of Breach </th>
+    <th> Priority </th>
+</tr>
+{% for row in output %}
+    <tr>
+      <td>{{ row[0]}}</td>
+      <td>{{ row[6]}}</td>
+      <td>{{ row[7]}}</td>
+      <td>{{ row[8]}}</td>
+    </tr>
+{% endfor %}
+</table>
+</br>
+
+{% elif show_one %}
+<title>Breach Details </title>
+<table width = "500px">
+<tr>
+    <th> ID </th>
+    <th> Affected Organization </th>
+    <th> STIX Output </th>
+</tr>
+    <tr>
+      <td>{{ output[0]}}</td>
+      <td>{{ output[5]}}</td>
+      <td><a href = {{url_for('breach_results', breach_id=output[0] ,_external = True, format="stix")}} > download</a> </td>
+    </tr>
+</table>
+</br>
+{% endif %}
+</div>
+{% endblock %}
