--- templates/display.html
@@ -1,152 +1,0 @@
-
-<!DOCTYPE html>
-
-<header>
-<link href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css" rel="stylesheet">
-
-    <div class="navbar navbar-inverse navbar-fixed-top" role="navigation">
-      <div class="container">
-        <div class="navbar-header">
-          <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-collapse">
-            <span class="sr-only">Toggle navigation</span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-            <span class="icon-bar"></span>
-          </button>
-          <a class="navbar-brand" href="/">Breach Submission System</a>
-        </div>
-        <div class="collapse navbar-collapse">
-          <ul class="nav navbar-nav">
-            <li class="active"><a href="/">Home</a></li>
-            <li><a href="/breach/list">Prior Submissions</a></li>
-            <li><a href="/about">About</a></li>
-          </ul>
-        </div><!--/.nav-collapse -->
-      </div>
-    </div>
-
-</header>
-
-
-<div class="container">
-
-</br>
-</br>
-</br>
-{% if breach_id %}
-<title>Success </title>
-<h1>Top men are on the case. 
-</h1>
-
-<h2>
-Refer to 
-<a href = {{url_for('breach_results', breach_id=breach_id ,_external = True)}} > this link</a> 
-to track the status of your submission 
-</h2>
-
-
-{% elif about %}
-<title>About Us </title>
-
-<p>Project created by Ben Schmoker <a href='http://github.com/bschmoker'>Follow him on Github</a> </p>
-
-{% elif show_all %}
-<title>Breach Details </title>
-<table width = "500px">
-<tr>
-    <th> ID </th>
-    <th> Timestamp </th>
-    <th> Type of Breach </th>
-    <th> Priority </th>
-</tr>
-{% for row in output %}
-    <tr>
-      <td>{{ row[0]}}</td> 
-      <td>{{ row[6]}}</td> 
-      <td>{{ row[7]}}</td> 
-      <td>{{ row[8]}}</td> 
-    </tr>
-{% endfor %}
-</table>
-</br>
-
-{% elif show_one %}
-<title>Breach Details </title>
-<table width = "500px">
-<tr>
-    <th> ID </th>
-    <th> Affected Organization </th>
-    <th> STIX Output </th>
-</tr>
-    <tr>
-      <td>{{ output[0]}}</td> 
-      <td>{{ output[5]}}</td> 
-      <td><a href = {{url_for('breach_results', breach_id=output[0] ,_external = True, format="stix")}} > download</a> </td>
-    </tr>
-</table>
-</br>
-
-{% else %}
-
-<title>Breach Submission </title>
-    <form method="POST" action="/breach/new" >
-    <p>
-    </br>
-    First, what is your name? <input type="text" name="submitter" required> </input>
-    </p>
-
-    <p>
-    How would you describe what happened? </br> <textarea maxlength = 100 rows="7" cols = "30" name="description" required> </textarea>
-    </p>
-
-    <p>
-    Who did it happen to?<input type="text" name="organization"required>
-    </p>
-
-    <p>
-<!--# TODO handle if blank, default empty string or None?  -->
-    What did they take (if anything) </br> <input type="text" name="asset" > </input>
-    </p>
-
-    <p>
-    What kind of damage did it cause?
-     <select name="damage">
-    <option value="Brand or Image Degradation">It damaged their image </option>
-    <option value="Loss of Competitive Advantage">They lost business to rivals </option>
-    <option value="Data Breach or Compromise" selected> Their data was stolen</option>
-    <option value="Regulatory, Compliance or Legal Impact">They might get sued</option>
-    <option value="Financial Loss"> They lost money</option>
-    <option value="Degradation of Service"> Their servers went down</option>
-     </select>
-    </p>
-
-
-    <p> When did it happen? 
-    <input type="date" name="timestamp" required> </input>
-    </p>
-
-
-    <p>
-    How confident is your assertion?
-     <select name="confidence">
-    <option value=High>Very Confident<option>
-    <option selected value=Medium>Sorta Confident</option>
-    <option value=Low>Not Confident</option>
-    <option value=Unknown>Just a rumor</option>
-     </select>
-    </p>
-
-    <p>
-     Is this information sensitive?
-    <input type="hidden" value="False" name="sensitive" />
-    <input type="checkbox" name="sensitive" value="True" checked > (select if true)
-     </input>
-    </p>
-
-
-        <input type="submit" value="Submit">
-    </form>
-    {% endif %}
-
-    </div><!-- /.container -->
-
