+++ utils/dbUtils.py
@@ -1,0 +1,59 @@
+from sqlite3 import dbapi2 as sqlite3
+from sqlite3 import IntegrityError
+from flask import make_response, Flask, request, url_for, render_template, g, abort # 'g' is a magical database thingie
+import os
+
+app = Flask(__name__)
+app.config.update(dict(
+    DATABASE=os.path.join(app.root_path, 'mockingjay.db')
+    ))
+
+# ----- DATABASE CODE -----
+def connect_db():
+    """Connects to the specific database."""
+    rv = sqlite3.connect(app.config['DATABASE'])
+    rv.row_factory = sqlite3.Row
+    return rv
+
+
+def init_db():
+    """Initializes the database."""
+    with app.app_context():
+        db = get_db()
+        with app.open_resource('schema.sql', mode='r') as f:
+            db.cursor().executescript(f.read())
+        db.commit()
+
+def get_db():
+    """Opens a new database connection if there is none yet for the
+    current application context.
+    """
+    if not hasattr(g, 'sqlite_db'):
+        g.sqlite_db = connect_db()
+    return g.sqlite_db
+
+def insert_db(table, fields=(), values=()):
+    db = get_db()
+    cur = db.cursor()
+    query = 'INSERT INTO %s (%s) VALUES (%s)' % (
+        table,
+        ', '.join(fields),
+        ', '.join(['?'] * len(values))
+    )
+    cur.execute(query, values)
+    db.commit()
+    rowId = cur.lastrowid
+    cur.close()
+    return rowId
+
+def query_db(query, args=(), one=False):
+    cur = get_db().execute(query, args)
+    rv = cur.fetchall()
+    cur.close()
+    return (rv[0] if rv else None) if one else rv
+
+@app.teardown_appcontext
+def close_db(error):
+    """Closes the database again at the end of the request."""
+    if hasattr(g, 'sqlite_db'):
+        g.sqlite_db.close()
